src2/requirements.txt
pygame==2.5.2
pyyaml==6.0.1
numpy==1.24.3 